+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/viewmodel/MessageListViewModelBinding.kt
@@ -45,13 +45,11 @@ import io.getstream.chat.android.ui.message.list.MessageListView
  *
  * @param view The [MessageListView] to bind the ViewModel to.
  * @param lifecycleOwner Current owner of the lifecycle in which the events are handled.
- * @param enforceUniqueReactions If message reactions are unique or a single user can post multiple reactions.
  */
 @JvmName("bind")
 public fun MessageListViewModel.bindView(
     view: MessageListView,
     lifecycleOwner: LifecycleOwner,
-    enforceUniqueReactions: Boolean = true,
 ) {
 
     deletedMessageVisibility.observe(lifecycleOwner) {
@@ -74,7 +72,7 @@ public fun MessageListViewModel.bindView(
     }
     view.setMessageRetryHandler { onEvent(RetryMessage(it)) }
     view.setMessageReactionHandler { message, reactionType ->
-        onEvent(MessageReaction(message, reactionType, enforceUnique = enforceUniqueReactions))
+        onEvent(MessageReaction(message, reactionType))
     }
     view.setMessageReplyHandler { cid, message -> onEvent(ReplyMessage(cid, message)) }
     view.setAttachmentDownloadHandler { downloadAttachmentCall ->